+++ __tests__/utils/createOmnichannelMessage.spec.ts
@@ -5,7 +5,7 @@ import { DeliveryMode, MessageType } from '../../src/core/messaging/OmnichannelM
 import PersonType from '@microsoft/omnichannel-ic3core/lib/model/PersonType';
 
 describe('createOmnichannelMessage', () => {
-    it ('createOmnichannelMessage with LiveChatV2 messaging contracts should return OmnichannelMessage contracts', () => {
+    it('createOmnichannelMessage with LiveChatV2 messaging contracts should return OmnichannelMessage contracts', () => {
         const amsReferences = ['id'];
         const amsMetadata = [{fileName: 'fileName.ext', size: 0, contentType: 'type'}]
         const sampleMessage = {
@@ -44,6 +44,74 @@ describe('createOmnichannelMessage', () => {
         expect(omnichannelMessage.fileMetadata).toEqual({
             fileSharingProtocolType: 0,
             id: amsReferences[0],
+            name: amsMetadata[0].fileName,
+            size: 0,
+            type: 'ext',
+            url: ''
+        });
+    });
+
+    it('createOmnichannelMessage with LiveChatV2 message containing \'amsreferences\' should return the correct fileMetadata', () => {
+        const amsReferences = ['id'];
+        const amsMetadata = [{fileName: 'fileName.ext', size: 0, contentType: 'type'}]
+        const sampleMessage = {
+            id: 'id',
+            content: '',
+            metadata: {
+                tags: 'tags',
+                amsMetadata: JSON.stringify(amsMetadata),
+                amsreferences: JSON.stringify(amsReferences)
+            },
+            sender: {
+                communicationUserId: 'id',
+                kind: "communicationUser"
+            },
+            senderDisplayName: 'senderDisplayName',
+            createdOn: 'createdOn'
+        };
+
+        const omnichannelMessage = createOmnichannelMessage(sampleMessage as any, {
+            liveChatVersion: LiveChatVersion.V2
+        });
+
+        expect(omnichannelMessage.fileMetadata).toEqual({
+            fileSharingProtocolType: 0,
+            id: amsReferences[0],
+            name: amsMetadata[0].fileName,
+            size: 0,
+            type: 'ext',
+            url: ''
+        });
+    });
+
+    it('createOmnichannelMessage with LiveChatV2 message should take precedence of \'amsreferences\'', () => {
+        const amsReferences = ['amsReferences'];
+        const amsreferences = ['amsreferences'];
+        const amsMetadata = [{fileName: 'fileName.ext', size: 0, contentType: 'type'}]
+        const sampleMessage = {
+            id: 'id',
+            content: '',
+            metadata: {
+                tags: 'tags',
+                amsMetadata: JSON.stringify(amsMetadata),
+                amsReferences: JSON.stringify(amsReferences),
+                amsreferences: JSON.stringify(amsreferences)
+            },
+            sender: {
+                communicationUserId: 'id',
+                kind: "communicationUser"
+            },
+            senderDisplayName: 'senderDisplayName',
+            createdOn: 'createdOn'
+        };
+
+        const omnichannelMessage = createOmnichannelMessage(sampleMessage as any, {
+            liveChatVersion: LiveChatVersion.V2
+        });
+
+        expect(omnichannelMessage.fileMetadata).toEqual({
+            fileSharingProtocolType: 0,
+            id: amsreferences[0],
             name: amsMetadata[0].fileName,
             size: 0,
             type: 'ext',