+++ CHANGELOG.md
@@ -5,9 +5,21 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Add `sendDefaultInitContext` optional parameter to `ChatSDK.startChat()` to automatically populate `browser`, `device`, `originurl` & `os` as default init context on web
 - Add `sendCacheHeaders` as optional paramater to `ChatSDK.initialize()` and `ChatSDK.getLiveChatConfig()`
+- Add `validateAuthChatRecord` call on `ChatSDK.startChat()` with `liveChatContext` for all authenticated chat scenarios
+- Pass `ChatClient` during `ACSAdapter` initialization
+- Pass `multiClient` to `AMSClient` on initialization to support `ChatSDK` multi-client
+
+### Fixed
+- Prevent `AMSFileManager.getFileIds()` & `AMSFileManager.getFileMetadata()` to be triggered on all activities with null checks
+- Add `LiveChatVersion` check on `ChatSDK.updateChatToken()`
+- Use `amsreferences` property instead of `amsReferences` by default
 
 ### Changed
 - Uptake [@microsoft/ocsdk@0.3.1](https://www.npmjs.com/package/@microsoft/ocsdk/v/0.3.1)
+- Uptake [acs_webchat-chat-adapter@0.0.35-beta.8](https://unpkg.com/acs_webchat-chat-adapter@0.0.35-beta.8/dist/chat-adapter.js)
+- Uptake [acs_webchat-chat-adapter@0.0.35-beta.9](https://unpkg.com/acs_webchat-chat-adapter@0.0.35-beta.9/dist/chat-adapter.js)
+- Uptake [acs_webchat-chat-adapter@0.0.35-beta.12](https://unpkg.com/acs_webchat-chat-adapter@0.0.35-beta.12/dist/chat-adapter.js)
+- Uptake [@microsoft/omnichannel-amsclient@0.1.2](https://www.npmjs.com/package/@microsoft/omnichannel-amsclient/v/0.1.2)
 
 ## [1.1.0] - 2022-04-15
 ### Added
@@ -26,7 +38,7 @@ All notable changes to this project will be documented in this file.
 - Add support for separate bot post chat survey feature
 - Pass `logger` to `acs_webchat-chat-adapter`
 
-### Fix
+### Fixed
 - Add `acs_webchat-chat-adapter` middlewares to format `channelData.tags`
 - Skip `session init` call on existing conversation
 - Fix `chat reconnect` not ending the conversation on calling `ChatSDK.endChat()`